Commit automatico

+++ src/main/java/br/com/sysloccOficial/financeiro/dao/RelatorioEventoDAO.java
@@ -341,14 +341,15 @@ public class RelatorioEventoDAO {
 	public void salvaCacheDoEvento(RelatorioEventos relatorioEvento){
 		
 		
-		/*	TypedQuery<CacheEvento> cacheTeste = manager.createQuery("SELECT c FROM CacheEvento c WHERE relatorioEvento="+relatorioEvento.getIdRelatorioEvento(), CacheEvento.class);
+		TypedQuery<CacheEvento> cacheTeste = manager.createQuery("SELECT c FROM CacheEvento c WHERE relatorioEvento="+relatorioEvento.getIdRelatorioEvento(), CacheEvento.class);
 		List<CacheEvento> caches = cacheTeste.getResultList();
-		*/
 		
-		
-		manager.createQuery("DELETE FROM CacheEvento WHERE relatorioEvento="+relatorioEvento.getIdRelatorioEvento()).executeUpdate();
-		manager.clear();
-		
+		if(caches.isEmpty() || caches.equals(null)){
+			
+		}else{
+			manager.createQuery("DELETE FROM CacheEvento WHERE relatorioEvento="+relatorioEvento.getIdRelatorioEvento()).executeUpdate();
+			manager.clear();
+		}
 		
 		List<CachePadrao> cachePadrao =  cacheEvento.listaRelatorioCaches(relatorioEvento.getIdLista());
